+++ lib/collaboration_web.ex
@@ -75,16 +75,6 @@ defmodule CollaborationWeb do
     end
   end
 
-  def channel do
-    quote do
-      use Phoenix.Channel
-      import CollaborationWeb.Gettext
-      import CollaborationWeb.ErrorHelpers
-
-      import CollaborationWeb.UserSocket, only: [user?: 1]
-    end
-  end
-
   @doc """
   When used, dispatch to the appropriate controller/view/etc.
   """