+++ lib/collaboration_web/controllers/comment_controller.ex
@@ -8,10 +8,20 @@ defmodule CollaborationWeb.CommentController do
         conn
         |> redirect(to: Routes.topic_path(conn, :show))
 
-      {:error, changeset} ->
+      {:error, _changeset} ->
         conn
         |> put_flash(:error, "Comment could not be created.")
         |> redirect(to: Routes.topic_path(conn, :show))
     end
+  end
+
+  def like(conn, %{"comment_id" => comment_id }) do
+    like_comment conn.assigns.current_user, comment_id
+    redirect conn, to: Routes.topic_path(conn, :show)
+  end
+
+  def unlike(conn, %{"comment_id" => comment_id }) do
+    unlike_comment conn.assigns.current_user, comment_id
+    redirect conn, to: Routes.topic_path(conn, :show)
   end
 end