+++ lib/collaboration_web/router.ex
@@ -49,7 +49,6 @@ defmodule CollaborationWeb.Router do
     # add protected resources below
     resources "/topics", TopicController,
       only: [:index, :new, :create, :edit, :update]
-    delete "/", IdeaController, :delete
     resources "/users",  UserController, only: [:index, :update]
     post "/feature/:id", TopicController, :feature
     get "/participants", UserController, :participants
@@ -60,6 +59,10 @@ defmodule CollaborationWeb.Router do
 
     # add protected resources below
     resources "/comments",  CommentController, only: [:create]
+    post "/rate/:idea_id/:rating", IdeaController, :rate
+    delete "/rate/:idea_id", IdeaController, :unrate
+    post "/like/:comment_id", CommentController, :like
+    delete "/like/:comment_id", CommentController, :unlike
 
     post "/", IdeaController, :create
     post "/complete", UserController, :finish